NetVarHooks cleanup
