+++ Fedoraware/TeamFortress2/TeamFortress2/Features/NetVarHooks/NetVarHooks.h
@@ -1,15 +1,13 @@
 #pragma once
 #include "../Feature.h"
 
-#include "../../Hooks/HookManager.h"
 #include "../../SDK/Includes/proxyfnhook.h"
-#include "../Backtrack/Backtrack.h"
 #include "../Vars.h"
 
 class NetVarHooks
 {
 private:
-	bool HookNetvar(std::vector<std::string> path, ProxyFnHook& hook, RecvVarProxyFn function);
+	bool HookNetvar(const std::vector<std::string>& path, ProxyFnHook& hook, RecvVarProxyFn function);
 
 public:
 	void Init();
